+++ FeedForwardDriver.cpp
@@ -3,28 +3,33 @@
 
 int main()
 {
-    TLDEDA001::AN_Network Network(3, 2, 0.01, 0, 10,true);
+    //Network instantiated with 3 inputs, 2 hidden layer nodes, 1 output layer node (no user input),
+    // a learning rate of 0.01, threshold of 0 (unnecessary in this example), 10 epochs, uses sigmoid activation funtion
+    TLDEDA001::AN_Network Network(3, 2, 0.01, 0, 10, true);
 
-    std::vector<float> inputs;
+    //given inputs
+    std::vector<double> inputs;
     inputs.push_back(1.30);
     inputs.push_back(2.70);
     inputs.push_back(0.80);
 
+    //given biases
     Network.getNode(0, 0).setBias(0.1);
     Network.getNode(0, 1).setBias(-0.3);
     Network.getNode(1, 0).setBias(-0.3);
 
-    std::vector<float> weights1;
+    //given weights
+    std::vector<double> weights1;
     weights1.push_back(0.1);
     weights1.push_back(0.2);
     weights1.push_back(0.5);
 
-    std::vector<float> weights2;
+    std::vector<double> weights2;
     weights2.push_back(-0.4);
     weights2.push_back(1.0);
     weights2.push_back(-0.6);
 
-    std::vector<float> weights3;
+    std::vector<double> weights3;
     weights3.push_back(0.8);
     weights3.push_back(1.0);
 
@@ -32,6 +37,7 @@ int main()
     Network.getNode(0, 1).setWeights(weights2);
     Network.getNode(1, 0).setWeights(weights3);
 
+    //Answers
     std::cout << "Q3: Hidden layer nodes outputs:" << std::endl;
     std::cout << Network.PredictIndividual(0, 0, inputs) << std::endl;
     std::cout << Network.PredictIndividual(0, 1, inputs) << std::endl;
