+++ TrainingData.h
@@ -11,22 +11,22 @@ namespace TLDEDA001
     class TrainingData
     {
         //list of inputs and outputs
-        std::vector<float> input1;
-        std::vector<float> input2;
-        std::vector<float> output;
+        std::vector<double> input1;
+        std::vector<double> input2;
+        std::vector<double> output;
 
     public:
         //Add a record
-        void AddDataSet(const float input1, const float input2, const float output);
+        void AddDataSet(const double input1, const double input2, const double output);
 
         //return the first input at specifed row index
-        const float getInput1(int index) const;
+        const double getInput1(int index) const;
         //return the second input at specifed row index
-        const float getInput2(int index) const;
+        const double getInput2(int index) const;
         //return the output at specifed row index
-        const float getOutput(int index) const;
+        const double getOutput(int index) const;
         //Return input1 and input2 in a const vector
-        const std::vector<float> getInputs(const int index) const;
+        const std::vector<double> getInputs(const int index) const;
         //get number of records
         const int getSize() const;
         //clear all data/records
